Port to 1.12.5, Forge version does not work for now
1. The Mobz shield can now be repaired just like the vanilla one
2. Keep using old spawn egg texture
3. The old vanilla spawn egg texture and model are embedded
4. appendHoverText() are moved to Items and BlockItems
5. equipment assets from 'models/equipment' to 'equipment'
6. The fabric version no longer depend on FabricShieldLib
7. TODO: Fix pigs texture

+++ src/main/java/net/mobz/entity/FrostBlaze.java
@@ -81,7 +81,7 @@ public class FrostBlaze extends BlazeLike {
 	}
 
 	@Override
-	public boolean causeFallDamage(float float_1, float float_2, DamageSource source) {
+	public boolean causeFallDamage(double float_1, float float_2, DamageSource source) {
 		return false;
 	}
